refactor: data context 파일 응집
파일이 좀 과하게 분리되어 있는 것 같아 응집시킴
