--- src/context/DataStore/constants.ts
@@ -1,14 +1,0 @@
-import { DataStore } from '../../types/dataStore';
-
-export const initialState: DataStore = {
-  products: {
-    data: null,
-    loading: false,
-    error: null,
-  },
-  'cart-items': {
-    data: null,
-    loading: false,
-    error: null,
-  },
-};
