+++ src/context/DataStore/useDataContext.ts
@@ -1,13 +1,12 @@
 import { useContext } from 'react';
-import { DataContext, DataContextValue } from './DataContext';
+import { DataContext } from './DataProvider';
 
-export const useDataContext = (): DataContextValue => {
+export const useDataContext = () => {
   const context = useContext(DataContext);
 
   if (context === undefined) {
     throw new Error(
-      'useDataContext는 반드시 DataProvider 내에서 사용되어야 합니다. ' +
-        '컴포넌트가 <DataProvider>로 감싸져 있는지 확인하세요.',
+      'useDataContext는 반드시 DataProvider 내에서 사용되어야 합니다. 컴포넌트가 <DataProvider>로 감싸져 있는지 확인하세요.',
     );
   }
